add exception handler for rest of models

+++ src/main/java/kapsalon/nl/services/AppointmentServiceImpl.java
@@ -1,5 +1,7 @@
 package kapsalon.nl.services;
 import jakarta.persistence.EntityNotFoundException;
+import kapsalon.nl.exceptions.ErrorMessages;
+import kapsalon.nl.exceptions.RecordNotFoundException;
 import kapsalon.nl.models.dto.AppointmentDTO;
 import kapsalon.nl.models.entity.*;
 import kapsalon.nl.repo.*;
@@ -36,12 +38,12 @@ public class AppointmentServiceImpl implements AppointmentService {
 
     @Override
     public AppointmentDTO getAppointmentById(Long id) {
-
         Optional<Appointment> entity = appointmentRepository.findById(id);
-        AppointmentDTO dto;
-        dto = fromEntityToDto(entity.get());
-
-        return dto;
+        if (entity.isPresent()) {
+            return fromEntityToDto(entity.get());
+        }else {
+            throw new RecordNotFoundException(ErrorMessages.APPOINTMENT_NOT_FOUND + id);
+        }
     }
 
     @Override
@@ -143,7 +145,13 @@ public class AppointmentServiceImpl implements AppointmentService {
 
     @Override
     public AppointmentDTO deleteAppointment(Long id) {
-        appointmentRepository.deleteById(id);
+        Optional<Appointment> entity = appointmentRepository.findById(id);
+        if (entity.isPresent()) {
+            appointmentRepository.deleteById(id);
+        }else {
+
+            throw new RecordNotFoundException(ErrorMessages.BARBER_NOT_FOUND + id);
+        }
         return null;
     }
